working on splash styling
