+++ react-app/src/components/SplashPage/appstore.css
@@ -1,18 +1,11 @@
-@@ -0,0 +1,51 @@
 @import url(https://fonts.googleapis.com/css?family=Lato:400,700,300,100);
 @import url(https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css);
 @import url(https://kit.fontawesome.com/a076d05399.js);
-.wrapper {
-    margin-top: 10rem;
-    margin-left: 7rem;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    /* background: rgb(21, 23, 24); */
-}
 
-.app-store {
+.appStore__buttonContainer {
+    display: inline-block;
+}
+.appStore__button {
     background: rgb(21, 23, 24);
     border: 1px solid white;
     border-radius: 2px;
@@ -21,32 +14,30 @@
     position: relative;
     text-align: left;
     transition: transform 0.25s, z-index 0s 0.25s;
-    width: 170px;
+    width: 17rem;
     z-index: 1;
     overflow: hidden;
 }
-.app-store:hover {
+.appStore__button:hover {
     transform: scale(1.2);
     transition: transform 0.25s;
     z-index: 2;
     background: rgb(21, 23, 24);
 }
-.app-store .app-store-icon {
+.appStore__button .appStore__icon {
     background: rgb(21, 23, 24);
-
     left: 0.7em;
     font-size: 2rem;
     top: 50%;
     transform: translate(-50%, -50%);
     position: absolute;
 }
-.app-store p {
+.appStore__button p {
     font-size: 0.75rem;
     background: rgb(21, 23, 24);
     margin-bottom: -0.5em;
 }
-.app-store h1 {
+.appStore__button h1 {
     font-size: 1.5rem;
-
     background: rgb(21, 23, 24);
 }
